+++ src/components/Game/GameContent/GameContentPlayField/EatPlayField.vue
@@ -6,7 +6,7 @@
         </div>
         <div class="row">
             <div class="col-12">
-                <h3 class="text-center">
+                <h3 id="eat-play-field-action-text" class="text-center">
                     <VRoller :default-char="eatTargetText" :text="eatTargetText" class="d-inline-flex"/>
                     <CancelPlayerTarget :play="play" attribute="eaten" class="ml-2" @playerSelected="playerSelected"/>
                 </h3>
@@ -53,6 +53,11 @@ export default {
 };
 </script>
 
-<style scoped>
+<style lang="scss" scoped>
+    @import "../../../../../node_modules/bootstrap/scss/bootstrap";
+    @import "../../../../assets/scss/variables";
 
+    #eat-play-field-action-text {
+        @include font-size(1.5rem);
+    }
 </style>