+++ src/components/GameLobby/GameLobbyAlreadyHavePlayingGame.vue
@@ -61,9 +61,9 @@ export default {
     computed: {
         gamePhaseText() {
             if (this.game.phase === "day") {
-                return `<i class="fas fa-sun mr-2"></i>${this.$t("GameLobbyAlreadyHavePlayingGame.day")} ${this.game.turn}`;
+                return `<i class="fas fa-sun sun-color mr-2"></i>${this.$t("GameLobbyAlreadyHavePlayingGame.day")} ${this.game.turn}`;
             } else {
-                return `<i class="fas fa-moon mr-2"></i>${this.$t("GameLobbyAlreadyHavePlayingGame.night")} ${this.game.turn}`;
+                return `<i class="fas fa-moon moon-color mr-2"></i>${this.$t("GameLobbyAlreadyHavePlayingGame.night")} ${this.game.turn}`;
             }
         },
     },